Return int, not WORD, for error codes and the like
Also, drop declarations of some non-existing functions, and
change a few (but not all) of the functions that always succeed
to void.

+++ sources/compiler.c
@@ -2243,7 +2243,7 @@ dopowerd:
 
 WORD GenerateFactors(WORD n,WORD inc)
 {
-	WORD subexp;
+	int subexp;
 	int i, error = 0;
 	SBYTE *s;
 	SBYTE *tokenbuffer = (SBYTE *)Malloc1(8*n*sizeof(SBYTE),"GenerateFactors");
